refactor(backend): modularize AuthService into partial domain modules
